+++ pages/detail/book/bookDetail.tsx
@@ -1,4 +1,3 @@
-import Head from "next/head";
 import BookDetailComponent from "../../../src/components/detail/book/BookDetail";
 
 export default function BookDetail(): JSX.Element {
