+++ src/components/detail/book/BookDetail.tsx
@@ -1,7 +1,8 @@
-import {BookCover, BookInfo, TabContainer} from "../../../containers/detail/book/bookDetailContainer";
+import {BookCover, BookInfo} from "../../../containers/detail/book/bookDetailContainer";
 import {DivLine} from "../../../styles/detail/book/bookDetailStyle";
 import {NewStudyWrapper, StudyBackground} from "../../../styles/main/BooksStyle";
 import {MoreButton, StudyContainer, ThisBookBox} from "../../../containers/main/BooksContainer";
+import {TabContainer} from "../../../containers/detail/common/commonContainer";
 
 export default function BookDetailComponent(): JSX.Element {
   return (
