+++ src/styles/detail/book/bookDetailStyle.tsx
@@ -9,6 +9,7 @@ const MainBookCover = styled.div`
   align-items: center;
 `;
 
+
 const BookImg = styled.img`
   width: 138px;
   height: 200px;
@@ -27,45 +28,9 @@ const TitleInner = styled.div`
   padding: 50px 87px 0 20px;
 `;
 
-const Title = styled.div`
-  font-size: 32px;
-  color: #222222;
-  font-weight: bold;
-  line-height: 40px;
-
-`;
-
-const BookDetailWrapper = styled.div`
-  display: flex;
-  justify-content: space-around;
-  padding-top: 10px
-`;
-
-const BookDetail = styled.div`
-  font-size: 14px;
-  color: #777777;
-  padding-right: 10px;
-  border-right-color: #DDDDDD;
-`;
-
 const DivLine = styled.div`
   height: 10px;
   background-color: #F1F1F3;  
-`;
-
-const TabWrapper = styled.div`
-  display: flex;
-  justify-content: space-around;
-  height: 60px;
-`;
-
-const TabButton = styled.div<{active: boolean}>`
-  border-bottom: ${(props) => (props.active ? '1px solid #FFD262' : '1px solid #DDDDDD')};
-  width: 33%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
 `;
 
 const ThisBookContentWrapper = styled.div`
@@ -76,14 +41,9 @@ const ThisBookContentWrapper = styled.div`
 
 export {
     ThisBookContentWrapper,
-    TabWrapper,
-    TabButton,
     DivLine,
     TitleInner,
     MainBookCover,
     BookImg,
-    TitleInfoWrapper,
-    Title,
-    BookDetailWrapper,
-    BookDetail
+    TitleInfoWrapper
 }
